+++ examples/nest-graphql-typeorm/src/todo-item/types.ts
@@ -5,4 +5,4 @@ import { TodoItemDTO } from './dto/todo-item.dto';
 export const TodoItemConnection = ConnectionType(TodoItemDTO);
 
 @ArgsType()
-export class TodoItemQuery extends QueryArgsType(TodoItemDTO) {}
+export class TodoItemQuery extends QueryArgsType(TodoItemDTO, { defaultResultSize: 2 }) {}