+++ packages/query-graphql/src/types/query/query-args.type.ts
@@ -1,7 +1,8 @@
 import { Class, Filter, Query, SortField } from '@nestjs-query/core';
 import { ArgsType, Field } from 'type-graphql';
-import { ValidateNested } from 'class-validator';
+import { ValidateNested, Validate } from 'class-validator';
 import { Type } from 'class-transformer';
+import { PropertyMax } from '../validators/property-max.validator';
 import { FilterType } from './filter.type';
 import { SortType } from './sorting.type';
 import { CursorPagingType } from './paging.type';
@@ -16,7 +17,32 @@ export interface QueryArgsType<T> extends Query<T> {
   paging?: CursorPagingType;
 }
 
-export function QueryArgsType<T>(TClass: Class<T>): StaticQueryType<T> {
+export interface QueryArgsTypeOpts<T> {
+  /**
+   * The default number of results to return.
+   * [Default=10]
+   */
+  defaultResultSize?: number;
+  /**
+   * The maximum number of results that can be returned from a query.
+   * [Default=50]
+   */
+  maxResultsSize?: number;
+  /**
+   * The default sort for queries.
+   * [Default=[]]
+   */
+  defaultSort?: SortField<T>[];
+  /**
+   * Default filter.
+   * [Default=\{\}]
+   */
+  defaultFilter?: Filter<T>;
+}
+
+const defaultQueryOpts = { defaultResultSize: 10, maxResultsSize: 50, defaultSort: [], defaultFilter: {} };
+
+export function QueryArgsType<T>(TClass: Class<T>, opts: QueryArgsTypeOpts<T> = defaultQueryOpts): StaticQueryType<T> {
   const F = FilterType(TClass);
   const S = SortType(TClass);
   const P = CursorPagingType();
@@ -29,17 +55,28 @@ export function QueryArgsType<T>(TClass: Class<T>): StaticQueryType<T> {
 
     static PageType = P;
 
-    @Field(() => P, { defaultValue: new P(), description: 'Limit or page results' })
+    @Field(() => P, {
+      defaultValue: { first: opts.defaultResultSize ?? defaultQueryOpts.defaultResultSize },
+      description: 'Limit or page results.',
+    })
     @ValidateNested()
+    @Validate(PropertyMax, ['first', opts.maxResultsSize ?? defaultQueryOpts.maxResultsSize])
+    @Validate(PropertyMax, ['last', opts.maxResultsSize ?? defaultQueryOpts.maxResultsSize])
     @Type(() => P)
     paging?: CursorPagingType;
 
-    @Field(() => F, { defaultValue: new F(), description: 'Specify to filter the records returned.' })
+    @Field(() => F, {
+      defaultValue: opts.defaultFilter ?? defaultQueryOpts.defaultFilter,
+      description: 'Specify to filter the records returned.',
+    })
     @ValidateNested()
     @Type(() => F)
     filter?: Filter<T>;
 
-    @Field(() => [S], { defaultValue: [], description: 'Specify to sort results' })
+    @Field(() => [S], {
+      defaultValue: opts.defaultSort ?? defaultQueryOpts.defaultSort,
+      description: 'Specify to sort results.',
+    })
     @ValidateNested()
     @Type(() => S)
     sorting?: SortField<T>[];