+++ HISTORY.md
@@ -1,12 +1,23 @@
+# v0.0.5
+
+* Add ability to specify query defaults.
+   * `defaultResultSize` -  the default number of results to return from a query
+   * `maxResultsSize` -  the maximum number of results to return from a query
+   * `defaultSort` -  The default sort to apply to queries
+   * `defaultFilter` -  The default filter to apply to queries
+
+# v0.0.4
+
+* Add files field to `@nestjs-query/core` `package.json`
+
 # v0.0.3
 
-* Fix package readmes
+* Fix package READMEs
 * Add security scanning on sub modules.
 
 # v0.0.2
 
 * Add MIT license
-
 
 # v0.0.1